mostly fix upgrade to apollo-server v2
