+++ api/src/data/index.js
@@ -35,6 +35,8 @@ export const schema = [
       authenticate(identity: String!, password: String!): Authentication
       registerPerson(email: String!, password: String!): Authentication
     }
+
+    ${process.env.NODE_ENV === 'test' ? `scalar Upload` : null}
   `,
 ];
