+++ api/src/index.js
@@ -1,35 +1,8 @@
 import dotenv from 'dotenv/config'; // eslint-disable-line
 import { ApolloServer } from 'apollo-server';
-import { get } from 'lodash';
-import { resolvers, schema, models, dataSources } from './data';
+import { resolvers, schema, dataSources } from './data';
 
-import RockConnector from './connectors/rock';
-
-// Construct a context object for each request
-export const getContext = ({ req = {} } = {}) => {
-  // initialize connectors for every request so API fetches
-  // are deduplicated per-request only.
-  const connectors = {
-    Rock: new RockConnector(),
-  };
-
-  const initiatedModels = {};
-
-  const context = {
-    models: initiatedModels,
-    connectors,
-  };
-
-  Object.keys(models).forEach((modelName) => {
-    initiatedModels[modelName] = new models[modelName](context);
-  });
-
-  if (get(req, 'headers.authorization')) {
-    initiatedModels.Auth.registerToken(req.headers.authorization);
-  }
-
-  return context;
-};
+import getContext from './getContext';
 
 export default new ApolloServer({
   typeDefs: schema,
